+++ Boolti/Boolti/Sources/Entities/QR/QRScannerEntity.swift
