+++ Boolti/Boolti/Sources/Entities/Ticket/TicketDetailItemEntity.swift
