+++ Boolti/Boolti/Sources/Entities/Ticket/TicketItemEntity.swift
