+++ Boolti/Boolti/Sources/Entities/Ticket/TicketReservationDetailEntity.swift
