+++ Boolti/Boolti/Sources/Network/Repositories/Protocols/AuthRepositoryType.swift
