+++ Boolti/Boolti/Sources/Network/Repositories/Protocols/QRRepositoryType.swift
