+++ Boolti/Boolti/Sources/Network/Repositories/TicketReservationsRepository.swift
@@ -9,14 +9,7 @@ import Foundation
 
 import RxSwift
 
-protocol TicketReservationsRepositoryType {
-    var networkService: NetworkProviderType { get }
-    func ticketReservations() -> Single<[TicketReservationItemEntity]>
-    func ticketReservationDetail(with reservationID: String) -> Single<TicketReservationDetailEntity>
-    func requestRefund(with requestDTO: TicketRefundRequestDTO) -> Single<Void>
-}
-
-class TicketReservationRepository: TicketReservationsRepositoryType {
+final class TicketReservationRepository: TicketReservationsRepositoryType {
 
     let networkService: NetworkProviderType
