+++ Boolti/Boolti/Sources/UILayer/Concert/ConcertList/Cells/CheckingTicketCollectionViewCell.swift
