+++ Boolti/Boolti/Sources/UILayer/MyPage/QRScanner/QRScannerList/Cells/QRScannerListTableViewCell.swift
